move utils.nn into nn
