+++ gluoncv/nn/__init__.py
@@ -5,3 +5,5 @@ from . import bbox
 from . import coder
 from . import feature
 from . import predictor
+from . import matcher
+from . import sampler
