--- gluoncv/utils/nn/__init__.py
@@ -1,1 +1,0 @@
-"""GluonCV utility neural network layers."""
